requirements.txt
nextcord
openai
python-dotenv
easy_pil
pillow